app/requirements.txt
flask==3.0.3
torch>=2.0.0
transformers>=4.39.0
langchain>=0.1.13
langgraph>=0.4.1
langchain-community>=0.0.30
chromadb>=0.4.24
beautifulsoup4>=4.12.2
spacy>=3.7.2
langchain-text-splitters>=0.0.1
requests>=2.31.0
compressed-tensors>=0.9.0
accelerate>=1.3.0
huggingface-hub>=0.30.0
protobuf<=3.20.1
googlesearch-python>=1.3.0
selenium>=4.31.0